requirements.txt
ipywidgets
sidecar
matplotlib
mujoco
jupyterlab
matplotlib==3.7.5
notebook
ollama==0.5.1
open_clip_torch
openai_clip==1.0.1
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
owlready2==0.47
Pillow==11.3.0
Requests==2.32.4
rf_clip==1.1
torch==2.4.0
transformers==4.47.1
torchvision==0.19.0
jupyterlab_widgets
ipywidgets
ipyevents
pandas
ipython
sparqlwrapper
flask